rp2350: make target clean under CI flags (-Werror, Ninja, no picotool)
Verified with an exact CI-parity build (Ninja + WARNINGS_AS_ERRORS=ON +
RelWithDebInfo, matching ci.yml): RP2350_PICO previously failed 4 TUs with
-Werror. Fixes:

- fc/cli.c: re-enable the I2C/stack/config-size and ADC status blocks for
  RP2350 (removed the milestone-1 stub-era exclusion). RP2350 now has real
  I2C/ADC/config, so the previously excluded code compiles, links and reports
  real values; only the STM32/AT32 clock block stays excluded. The re-enable
  also makes the i2cErrorCounter declaration used on RP2350 (was unused-var).
- common/maths.h: INAV's type-safe single-eval MIN/MAX must win over the Pico
  SDK's guarded plain MIN/MAX (reached via build/atomic.h -> pico compiler.h
  before maths.h in fc_init.c). Undef first, so ordering no longer matters.
  No-op for every other target: nothing defines MIN/MAX before maths.h there.
- drivers/pwm_output_rp2350.c: sign-compare int maxTimServoCount vs 8u.
- cmake/rp2350.cmake: suppress -Wdouble-promotion for the vendored pico printf.c
  (per-file, mirroring stm32f7.cmake's vendor-HAL suppression); generate .uf2
  only when picotool is on PATH (GitHub runners don't ship it) — .hex/.bin are
  always produced and are what CI uploads.

CI-parity result: exit 0, zero warnings, .hex/.bin (and .uf2 locally)
produced. picotool is NOT needed in CI: apt on ubuntu-24.04 has it, but the
.uf2 step is conditional and CI only uploads .hex.

+++ src/main/drivers/pwm_output_rp2350.c
@@ -611,7 +611,7 @@ static void rp2350BuildOutputAssignment(timMotorServoHardware_t *out)
         const uint32_t flags = timerHardware[i].usageFlags;
         if (TIM_IS_MOTOR_ONLY(flags) && out->maxTimMotorCount < MAX_PWM_OUTPUTS) {
             out->timMotors[out->maxTimMotorCount++] = &timerHardware[i];
-        } else if (TIM_IS_SERVO_ONLY(flags) && out->maxTimServoCount < RP2350_MAX_SERVOS) {
+        } else if (TIM_IS_SERVO_ONLY(flags) && (uint)out->maxTimServoCount < RP2350_MAX_SERVOS) {
             out->timServos[out->maxTimServoCount++] = &timerHardware[i];
         }
     }
@@ -678,7 +678,7 @@ void pwmCalculateAssignment(timMotorServoHardware_t *out, const uint8_t *propose
             }
             motorIdx++;
         } else {
-            if (out->maxTimServoCount < RP2350_MAX_SERVOS && out->maxTimServoCount < MAX_PWM_OUTPUTS) {
+            if ((uint)out->maxTimServoCount < RP2350_MAX_SERVOS && out->maxTimServoCount < MAX_PWM_OUTPUTS) {
                 out->timServos[out->maxTimServoCount++] = timHw;
             }
         }
